Added new function getChampionData. Moved getAllChampionData and getAllChampionPagePaths into one file
